+++ NEWS.md
@@ -1,6 +1,7 @@
 # dev
 
 * feat: Support offset in learners `regr|classif.mgcv`, `regr.glm` and `regr.lmer`.
+* feat: Added learner `LearnerRegrQGam` and `LearnerRegrMQGam`
 
 # mlr3extralearners 1.0.0